Rename workflow run files
